+++ apps/admin/src/features/user/header/model/useGetUser.ts
@@ -1,6 +1,5 @@
+import { GET_USER } from '@/features/user/header/model/userQueries'
 import { useQuery } from '@apollo/client'
-
-import { GET_USER } from '../api/userQueries'
 
 export const useGetUser = (userId: number) => {
   const { data, error, loading } = useQuery(GET_USER, {
