+++ apps/admin/src/features/user/header/ui/userHeader.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/vars.scss';
+@import '../../../../styles/vars';
 
 .currentUserContainer {
   width: 100%;
