+++ apps/admin/src/features/user/uploadedPhotos/model/postsQuery.ts
