+++ apps/admin/src/features/user/uploadedPhotos/ui/uploadedPhotos.module.scss
